pyproject.toml
[project]
name = "scisummarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.10.9",
    "arxiv>=2.1.3",
    "baml-py>=0.59.0",
    "pydantic>=2.9.2",
    "pymupdf>=1.24.11",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
]
